+++ R/calculate_protein_abundance.R
@@ -1,7 +1,8 @@
 #' Label-free protein quantification
 #'
 #' Determines relative protein abundances from ion quantification. Only proteins with at least
-#' three peptides are considered for quantification.
+#' three peptides are considered for quantification. The three peptide rule applies for each
+#' sample independently.
 #'
 #' @param data a data frame that contains at least the input variables.
 #' @param sample a character column in the \code{data} data frame that contains the sample name.
@@ -123,7 +124,7 @@ calculate_protein_abundance <- function(data,
       {{ intensity_log2 }}
     ) %>%
     tidyr::drop_na() %>%
-    dplyr::group_by({{ protein_id }}) %>%
+    dplyr::group_by({{ protein_id }}, {{ sample }}) %>%
     dplyr::mutate(n_peptides = dplyr::n_distinct(!!rlang::ensym(peptide))) %>%
     dplyr::filter(.data$n_peptides > 2) %>%
     dplyr::select(-.data$n_peptides) %>%